requirements.txt
qdrant-client>=1.8.2
openai>=1.30.0
tiktoken>=0.7.0
python-dotenv>=1.0.1
fastapi>=0.111.0
uvicorn>=0.30.0
httpx>=0.27.0
